99-Upload-Project-Python/requirements.txt
flask
gunicorn
azure-identity
azure-storage-blob
opencensus-ext-azure
opencensus-ext-flask